externals/faiss-service/requirements.txt
Click==7.0
delorean==1.0.0
filelock==3.0.12
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.2
pytz==2019.3
Werkzeug==0.16.0
pymongo==3.10.1
mkl==2019.0
faiss-cpu==1.7.2
faiss-gpu==1.7.2
pytest==5.2.0
pytest-flask==0.15.0
pytest-ordering==0.6
protobuf==3.20.*
grpcio-tools
grpcio